// src/app/modules/student/student.interface.ts
import { Model } from "mongoose";

export type TGuardian = {
    fatherName:string;
    fatherOccupation: string;
    fatherContactNo: string;
    motherName:string;
    motherOccupation: string;
    motherContactNo: string;
}

export type TUserName = {
    firstName: string;
    middleName?:string;
    lastName: string;
}

export type TLocalGuardian = {
    name: string;
    occupation: string;
    contactNo: string;
    address: string
}

export type TStudent = {
    id: string;
    password: string
    name: TUserName;
    gender: "male"|"female" | "other";
    dateOfBirth?: string;
    email: string;
    contactNo: string;
    emergencyContactNo: string;
    bloodGroup?: 'A+' | 'A-' | 'B+' | 'B-' | 'AB+' | 'AB-' | 'O+' | 'O-';
    presentAddress: string;
    permanentAddress: string;
    guardian: TGuardian;
    localGuardian: TLocalGuardian;
    profileImage?: string;
    isActive: "active" | "blocked";
    isDeleted: boolean;
  }


  //for creating static
  export interface StudentModel extends Model<TStudent> {
    isUserExists(id:string): Promise<TStudent | null>
  }



  //for creating instance method
//  export type StudentMethods = {
//     isUserExists(id: string) : Promise<TStudent | null>
//   }


  //model
//  export type StudentModel = Model<TStudent, Record<string, never>, StudentMethods>;